Open the target O_RDONLY for root for BTRFS_IOC_FILE_EXTENT_SAME.
Both in duperemove and btrfs-extent-same.

This partially fixes the ETXTBSY failure when deduping a live filesystem
that has executables someone might want to exec while we open the file.

Really, we want to check CAP_SYS_ADMIN rather than root, but I don't want
to introduce a dependency on libpcap for something that's hopefully short
lived: once it's fixed on the kernel side, this check should compare kernel
version.

Signed-off-by: Adam Borowski <kilobyte@angband.pl>